+++ packages/web/app/api/mutations/useAiPatientSummaryMutation.js
@@ -2,15 +2,25 @@ import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { useApi } from '../useApi';
 import { AI_PATIENT_SUMMARY_QUERY_KEY } from '../queries/useAiPatientSummaryQuery';
 
+// The POST/PUT endpoints return an aiDocument directly; the GET query holds a
+// `{ aiDocument, requiresRegeneration }` shape, so when a mutation succeeds we
+// write the new document straight into the query cache. This keeps the UI in
+// sync without waiting on a refetch and avoids stale data from prior mutations
+// being preferred over the latest server state.
+const writeAiDocumentToCache = (queryClient, patientId, aiDocument) => {
+  queryClient.setQueryData([AI_PATIENT_SUMMARY_QUERY_KEY, patientId], {
+    aiDocument,
+    requiresRegeneration: false,
+  });
+};
+
 export const useGenerateAiPatientSummary = patientId => {
   const api = useApi();
   const queryClient = useQueryClient();
 
   return useMutation({
     mutationFn: () => api.post(`ai/patient/summary/${encodeURIComponent(patientId)}`),
-    onSuccess: () => {
-      queryClient.invalidateQueries([AI_PATIENT_SUMMARY_QUERY_KEY, patientId]);
-    },
+    onSuccess: aiDocument => writeAiDocumentToCache(queryClient, patientId, aiDocument),
   });
 };
 
@@ -21,9 +31,7 @@ export const useSaveAiPatientSummary = patientId => {
   return useMutation({
     mutationFn: ({ id, content }) =>
       api.put(`ai/patient/summary/${encodeURIComponent(id)}`, { content }),
-    onSuccess: () => {
-      queryClient.invalidateQueries([AI_PATIENT_SUMMARY_QUERY_KEY, patientId]);
-    },
+    onSuccess: aiDocument => writeAiDocumentToCache(queryClient, patientId, aiDocument),
   });
 };
 
@@ -37,8 +45,6 @@ export const useDiscardAiPatientSummary = patientId => {
         content: null,
         status: 'discarded',
       }),
-    onSuccess: () => {
-      queryClient.invalidateQueries([AI_PATIENT_SUMMARY_QUERY_KEY, patientId]);
-    },
+    onSuccess: aiDocument => writeAiDocumentToCache(queryClient, patientId, aiDocument),
   });
 };
